requirements.txt
# Core ML
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Data
pandas>=2.0.0
regex>=2023.6.3
dateutil>=2.8.2
python-dateutil>=2.8.2
tqdm>=4.65.0
requests>=2.31.0

# Config & Experiment Tracking
hydra-core>=1.3.2
omegaconf>=2.3.0
mlflow>=2.6.0

# Serving API
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.3.0

# Windows .evtx parsing (optional; install on Windows only)
# evtx>=0.8.0

# Notebook / Dev
notebook>=7.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
